+++ R/weird.R
@@ -4,6 +4,8 @@
 #' @keywords internal
 "_PACKAGE"
 
+#' @importFrom stats quantile
+#' @importFrom evd fpot pgpd
 #' @importFrom ggplot2 autoplot labs ggplot geom_line geom_segment
 #' @importFrom ggplot2 geom_contour_filled geom_contour scale_fill_manual
 #' @export